Add display current upload URL in text field
Included the current upload URL for Production mode to the
StringFieldEditor set for this purpose.

Notice:
     - To set the value at the beginning (createFieldEditors), it has
to be set through the preference store. This is because after the
create method is called, there is another one (initialize) that calls
load on the fields. In particular, load retrieves the value from the
preference store