Add code to store current text in fields into preferences
Added code in the performOk method to store the input text set in the
string field editors for the upload URLs.

In this same line, removed the line where a value was being set to the
preference store at the moment of creation of the string field editors.

Notice:
     - This change removes the default value shown when first opening
eclipse (displays blank)