Add TestUploadURLModification and modify Prompter
Added a test to verify that upon change of values in the Preference
Store, the AuthenticationPrompter returns the same value as the one set
in it.

The AuthenticationPrompter no longer uses the value in the Messages
class but instead gets the value from the PrefsFacade.

NOTICE:
- For the case of testing the Test mode, it is necessary to add an
environment field TESTING_MODE (value 1) for it to pass. This makes the
other one fail (production mode)