Add TestModifyUploadUrl using SWTBot (UI)
Added a test to check that upon changes in the Preference Page
(CodingSpectator -> Uploading), particularly in the Test Upload URL, it
is reflected in the value returned by the Authentication Prompter.

It is set to be tested with Windows as OS.

TODO:
- Mac OS : SWTBot cannot access the “Eclipse” menu, because this is a
system menu, not from Eclipse. Therefore it is not directly possible to
test it with SWTBot (Even with the SWTBot Recorder).